Add lot of changes

+++ src/main/java/com/sopas/gallery/sopas_gallery/service/impl/CommentService.java
@@ -1,13 +1,18 @@
 package com.sopas.gallery.sopas_gallery.service.impl;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.AnonymousAuthenticationToken;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.sopas.gallery.sopas_gallery.dto.CommentDTO;
 import com.sopas.gallery.sopas_gallery.dto.Response;
 import com.sopas.gallery.sopas_gallery.entity.Comment;
+import com.sopas.gallery.sopas_gallery.entity.User;
 import com.sopas.gallery.sopas_gallery.exception.OurException;
 import com.sopas.gallery.sopas_gallery.repository.CommentRepository;
+import com.sopas.gallery.sopas_gallery.repository.UserRepository;
 import com.sopas.gallery.sopas_gallery.service.interfac.ICommentService;
 import com.sopas.gallery.sopas_gallery.utils.Utils;
 
@@ -16,6 +21,9 @@ import java.util.List;
 import jakarta.validation.ConstraintViolationException;
 
 public class CommentService implements ICommentService {
+
+    @Autowired
+    private UserRepository userRepository;
 
     @Autowired
     private CommentRepository commentRepository;
@@ -85,10 +93,19 @@ public class CommentService implements ICommentService {
 
     @Override
     public Response deleteComment(Long commentId) {
-Response response = new Response();
+        Response response = new Response();
 
         try {
-            commentRepository.findById(commentId).orElseThrow(()-> new OurException("Comment Not Found"));
+            Comment comment = commentRepository.findById(commentId).orElseThrow(()-> new OurException("Comment Not Found"));
+            
+            String authUsername = getCurrentUsername();
+
+            if(!comment.getUser().getUsername().equals(authUsername) && !isAdmin(authUsername)){
+                response.setStatusCode(403);
+                response.setMessage("You do not have permission to delete this comment.");
+                return response;
+            }
+            
             commentRepository.deleteById(commentId);
             response.setStatusCode(200);
             response.setMessage("Comment Successfully Deleted");
@@ -103,4 +120,57 @@ Response response = new Response();
         return response;
     }
 
+    @Override
+    public Response updateComment(Long commentId, String newContent) {
+        Response response = new Response();
+
+        try {
+            Comment comment = commentRepository.findById(commentId).orElseThrow(()-> new OurException("Comment Not Found"));
+            
+            String authUsername = getCurrentUsername();
+
+            if(!comment.getUser().getUsername().equals(authUsername) && !isAdmin(authUsername)){
+                response.setStatusCode(403);
+                response.setMessage("You do not have permission to modify this comment.");
+                return response;
+            }
+            if(newContent == null || newContent.isBlank()){
+                response.setStatusCode(400);
+                response.setMessage("You cannot leave your comment empty.");
+                return response;
+            }
+            comment.setContent(newContent);
+            
+            Comment updatedComment = commentRepository.save(comment);
+            CommentDTO commentDTO = Utils.mapCommentEntityToDTO(updatedComment);
+            response.setStatusCode(200);
+            response.setMessage("Comment Updated Successfully.");
+            response.setCommentDTO(commentDTO);
+        }catch(OurException e){
+            response.setStatusCode(404);
+            response.setMessage(e.getMessage());
+        }
+         catch (Exception e) {
+            response.setStatusCode(500);
+            response.setMessage("Error Updating the Image " + e.getMessage());
+        }
+        return response;
+    }
+
+
+     private String getCurrentUsername(){
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+    if (authentication instanceof AnonymousAuthenticationToken) {
+        return null;
+    }
+    return authentication.getName();
+    }
+
+        private boolean isAdmin(String username){
+            return userRepository.findByUsername(username)
+            .map(User::isAdmin)
+            .orElse(false);
+    }
+
+       
 }
